# server/swagger_spec.yaml
openapi: 3.0.3
info:
  title: KroomCare AI-Supported CRM API
  description: Dokumentasi API untuk platform KroomCare CRM, meliputi fitur Autentikasi (2FA), Tiket Bantuan, Poin & Loyalty, Forum Komunitas, dan Panel Dashboard Admin/Staf.
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Development Server
paths:
...
  /api/tickets:
    get:
      summary: Dapatkan Seluruh Daftar Tiket
      tags:
        - Support Tickets
      parameters:
        - name: role
          in: query
          required: false
          schema:
            type: string
        - name: userId
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Sukses mengambil daftar tiket
    post:
      summary: Buat Tiket Keluhan Baru
      tags:
        - Support Tickets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - title
                - category
                - description
              properties:
                user_id:
                  type: integer
                title:
                  type: string
                category:
                  type: string
                description:
                  type: string
                priority:
                  type: string
                  default: Low
      responses:
        '200':
          description: Tiket berhasil dibuat
...
  /api/forums:
    get:
      summary: Dapatkan Semua Thread Forum Komunitas
      tags:
        - Community Forum
      responses:
        '200':
          description: Sukses mengambil daftar thread
    post:
      summary: Buat Thread Diskusi Baru
      tags:
        - Community Forum
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - title
                - content
              properties:
                user_id:
                  type: integer
                title:
                  type: string
                content:
                  type: string
                tags:
                  type: string
      responses:
        '200':
          description: Thread baru berhasil diterbitkan
...